requirements.txt
asgiref==3.6.0
asyncpg==0.27.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
defusedxml==0.7.1
Django==3.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.8
django-filter==23.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.0.1
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
Pygments==2.16.1
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
whitenoise==6.5.0
